+++ NEWS.md
@@ -1,5 +1,8 @@
 # cea 0.0.0.9000
 
+* Added `CEAC()` to calculate cost-effectiveness acceptability curve from
+  regression model.
+
 * Added `QALYs()` and `Costs()` to extract fitted regression components.
 
 * Added `ci()` to calculate bootstrap confidence intervals from regression